LayoutTests/platform/glib/imported/w3c/web-platform-tests/navigation-timing/nav2_test_redirect_chain_xserver_final_original_origin-expected.txt

CONSOLE MESSAGE: Error: assert_equals: Expected redirectEnd to be 0. expected 0 but got 13.000000000000002
Description

This test validates the values of the window.performance.getEntriesByType("navigation")[0].redirectCount and the window.performance.getEntriesByType("navigation")[0].redirectStart/End times for a cross-origin server side redirect navigation when the final document is the same as the original origin.



FAIL Navigation Timing 2 WPT Error: assert_equals: Expected redirectEnd to be 0. expected 0 but got 13.000000000000002
